Break CLI off into own package
* so I can test it once i mock the dynamodb service
