+++ Makefile
@@ -5,7 +5,7 @@ PROJECT_NAME=confidynt ## @build Project name
 DEFAULT_TARGET=$(BUILD_DIR)/$(PROJECT_NAME)
 MAKEFILE_LIST = Makefile
 
-.PHONY: clean build install help help-variables
+.PHONY: clean build generate install help help-variables
 
 clean: ## @build Clean stuff
 		rm -rf $(BUILD_DIR)
@@ -16,6 +16,9 @@ build $(DEFAULT_TARGET): ## @build Build the actual thing
 		go build -o $(DEFAULT_TARGET)
 		# build for specific operating systems
 		BUILD_DIR=$(BUILD_DIR) PROJECT_NAME=$(PROJECT_NAME) ./build.sh "linux/amd64" "darwin/amd64"
+
+generate:
+	go generate ./...
 
 install: build ## @build Install make-doc to /usr/local/bin
 	sudo cp $(DEFAULT_TARGET) /usr/local/bin
