+++ types/config.go
@@ -1,0 +1,3 @@
+package types
+
+type Config map[string]string
